add some base case test
